requirements.txt
flask==3.0.2
pillow==10.2.0
opencv-python==4.9.0.80
numpy==1.26.4
openai==1.78.1
python-dotenv==1.1.0
requests==2.32.3 